+++ app/api/tags/controllers/tags.js
@@ -1,0 +1,17 @@
+"use strict";
+const { sanitizeEntity } = require("strapi-utils");
+
+/**
+ * Read the documentation (https://strapi.io/documentation/developer-docs/latest/development/backend-customization.html#core-controllers)
+ * to customize this controller
+ */
+
+module.exports = {
+  async find(ctx) {
+    const entities = await strapi.services.tags.find(ctx.query, []);
+
+    return entities.map((entity) =>
+      sanitizeEntity(entity, { model: strapi.models.tags })
+    );
+  },
+};