+++ app/extensions/users-permissions/controllers/User.js
@@ -11,7 +11,13 @@ const {
   getAverageStars,
   getSummaryDetails,
 } = require("../../../api/test/utils");
-const { sanitizeImage } = require("../../../utils/sanitizers");
+const {
+  sanitizeImage,
+  sanitizeOutcomes,
+  sanitizeLink,
+} = require("../../../utils/sanitizers");
+const { composeLink } = require("../../../utils/links");
+const { shuffleArray } = require("../../../utils/array");
 
 const sanitizeUser = (user) =>
   sanitizeEntity(user, {
@@ -23,7 +29,123 @@ const sanitizeAppointment = (appointment) =>
     model: strapi.query("appointment").model,
   });
 
+const ME_POPULATE = ["role", "tags"];
+
 module.exports = {
+  /**
+   * Retrieve authenticated user.
+   * @return {Object|Array}
+   */
+  async me(ctx) {
+    const id = ctx.state.user.id;
+    const userService = strapi.plugins["users-permissions"].services.user;
+
+    const user = await userService.fetch({ id }, ME_POPULATE);
+
+    if (!user) {
+      return ctx.badRequest(null, [
+        { messages: [{ id: "No authorization header was found" }] },
+      ]);
+    }
+
+    ctx.body = sanitizeUser(user);
+  },
+
+  /**
+   * Update user's tags
+   * @return {Object|Array}
+   */
+  async updateTags(ctx) {
+    const userId = ctx.state.user.id;
+    const { tags } = ctx.request.body;
+
+    const userService = strapi.plugins["users-permissions"].services.user;
+
+    await userService.edit({ id: userId }, { tags });
+
+    const user = await userService.fetch({ id: userId }, ME_POPULATE);
+
+    if (!user) {
+      return ctx.badRequest(null, [
+        { messages: [{ id: "No authorization header was found" }] },
+      ]);
+    }
+
+    ctx.body = sanitizeUser(user);
+  },
+
+  /**
+   * Retrieves interesting content based on user's tag preferences
+   * @return {Object|Array}
+   */
+  async interests(ctx) {
+    const userId = ctx.state.user.id;
+
+    const userService = strapi.plugins["users-permissions"].services.user;
+    const tagsService = strapi.services.tags;
+
+    const user = await userService.fetch({ id: userId }, [
+      ...ME_POPULATE,
+      "completed_tests",
+    ]);
+
+    if (!user) {
+      return ctx.badRequest(null, [
+        { messages: [{ id: "No authorization header was found" }] },
+      ]);
+    }
+
+    const completedTestIds = user.completed_tests?.map(
+      ({ test, test_snapshot }) => test ?? test_snapshot.id
+    );
+
+    const tagsIds = user.tags.map(({ id }) => id);
+
+    const tags = await tagsService.find({ id_in: tagsIds }, [
+      "pages.tags",
+      "tests.tags",
+    ]);
+
+    const uniquePages = {};
+    const uniqueTests = {};
+    tags.forEach(({ pages, tests }) => {
+      pages.forEach((page) => {
+        const { id, slug, title, lead, tags } = page;
+        uniquePages[id] = {
+          id,
+          slug,
+          title,
+          description: lead,
+          tags: tags.map(({ name }) => name),
+          type: "page",
+          link: sanitizeLink(composeLink("page")(page)),
+        };
+      });
+      tests.forEach((test) => {
+        const { id, slug, name, description, tags, type } = test;
+        // Include only tests that the user has not yet completed
+        if (!completedTestIds.includes(id)) {
+          uniqueTests[id] = {
+            id,
+            slug,
+            title: name,
+            description,
+            tags: tags.map(({ name }) => name),
+            type,
+            link: sanitizeLink(composeLink("test")(test)),
+          };
+        }
+      });
+    });
+
+    const pages = Object.keys(uniquePages).map((key) => uniquePages[key]);
+    const tests = Object.keys(uniqueTests).map((key) => uniqueTests[key]);
+
+    const interests = [...pages, ...tests];
+
+    ctx.body = shuffleArray(interests);
+  },
+
   async changePassword(ctx) {
     const ctxUser = ctx.state.user;
 
@@ -34,7 +156,7 @@ module.exports = {
     if (currentPassword && newPassword && newPassword === newPasswordConfirm) {
       const userService = strapi.plugins["users-permissions"].services.user;
 
-      const user = await userService.fetch({ id: ctxUser.id }, ["role"]);
+      const user = await userService.fetch({ id: ctxUser.id }, ME_POPULATE);
 
       if (!user) {
         return ctx.badRequest("User.change_password.user.not_exist");
@@ -51,12 +173,56 @@ module.exports = {
 
       // Note: New password will be hashed in userService's edit method below
       const updateData = { password: newPassword };
-      const data = await userService.edit({ id: user.id }, updateData);
+      await userService.edit({ id: user.id }, updateData);
+      const updatedUser = await userService.fetch({ id: user.id }, ME_POPULATE);
 
-      return ctx.send(sanitizeUser(data));
+      return ctx.send(sanitizeUser(updatedUser));
     }
 
     return ctx.badRequest("User.change_password.new_password.no_match");
+  },
+
+  async deleteAccount(ctx) {
+    const userId = ctx.state.user.id;
+
+    const userService = strapi.plugins["users-permissions"].services.user;
+    const completedTestsService = strapi.services["completed-tests"];
+    const appointmentService = strapi.services.appointment;
+
+    // Remove completed tests related to the user
+    const userTests = await completedTestsService.find({ user: userId });
+    if (userTests?.length) {
+      Promise.all(
+        userTests.map(async (test) => {
+          return await completedTestsService.delete({ id: test.id });
+        })
+      );
+    }
+
+    // Handle existing appointments related to user
+    const userAppointments = await appointmentService.find({ user: userId });
+    if (userAppointments?.length) {
+      Promise.all(
+        userAppointments.map(async (appointment) => {
+          return await appointmentService.update(
+            { id: appointment.id },
+            { status: "available", user: null }
+          );
+          // Nice to have: Send an email to the specialist to tell his/her appointment was "cancelled"
+        })
+      );
+    }
+
+    // Finally, delete user
+    const data = await userService.remove({ id: userId });
+
+    if (data?.id === userId) {
+      return {
+        message: "Account deleted successfully",
+      };
+    }
+
+    return ctx.badRequest("User.delete_account.fail");
   },
 
   async getSpecialist(id) {
@@ -138,7 +304,7 @@ module.exports = {
     );
 
     if (completed_test?.outcomes) {
-      ctx.body = completed_test?.outcomes;
+      ctx.body = await sanitizeOutcomes(completed_test.outcomes);
     }
   },